backend_chatbot/requirements.txt
Flask
python-dotenv
mysql-connector-python
google-generativeai
